Updated twitter-cldr classes for number formatter
