+++ lib/plurimath/formatter/number_data_reader.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Plurimath
+  module Formatter
+    class NumberDataReader < TwitterCldr::DataReaders::NumberDataReader
+      def formatter
+        NumberFormatter.new(self)
+      end
+    end
+  end
+end
