+++ lib/plurimath/formatter/numbers/integer.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-module TwitterCldr
-  module Formatters
+module Plurimath
+  module Formatter
     module Numbers
-      class Integer
+      class Integer < TwitterCldr::Formatters::Numbers::Integer
         def initialize(token, symbols = {})
           format     = token.value.split('.')[0]
           @format    = prepare_format(format, symbols)
